# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freelancer Platform API
  description: |
    Comprehensive API for a global freelance marketplace with secure payments, 
    verified profiles, milestone-based delivery, and dispute resolution.
  version: 1.0.0
  contact:
    name: NivixPe Team
    email: api@freelancerplatform.com
servers:
  - url: https://api.freelancerplatform.com/v1
    description: Production
  - url: https://staging-api.freelancerplatform.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, fullName, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                fullName:
                  type: string
                role:
                  type: string
                  enum: [FREELANCER, CLIENT]
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /disputes:
    post:
      tags: [Disputes]
      summary: File dispute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contractId, reason]
              properties:
                contractId:
                  type: string
                  format: uuid
                reason:
                  type: string
                evidence:
                  type: object
      responses:
        '201':
          description: Dispute filed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
  /disputes/{disputeId}:
    get:
      tags: [Disputes]
      summary: Get dispute by ID
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Dispute details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'

  # ==================== ADMIN ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [FREELANCER, CLIENT, ADMIN]
        fullName:
          type: string
        avatarUrl:
          type: string
        profileData:
          type: object
        rating:
          type: number
        totalReviews:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Dispute:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        filedBy:
          type: string
          format: uuid
        reason:
          type: string
        evidence:
          type: object
        status:
          type: string
          enum: [OPEN, UNDER_REVIEW, RESOLVED, CLOSED]
        resolvedBy:
          type: string
          format: uuid
        resolution:
          type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        statusCode:
          type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...